:recycle: Refactor actions
