+++ src/actions/constants.js
@@ -11,4 +11,9 @@ const MAPPINGS = {
 	MAPPINGS_FETCH_FAILURE: 'MAPPINGS_FETCH_FAILURE',
 };
 
-export { APP, MAPPINGS };
+const CELL = {
+	CELL_ACTIVE: 'CELL_ACTIVE',
+	CELL_HIGHLIGHT: 'CELL_HIGHLIGHT',
+};
+
+export { APP, CELL, MAPPINGS };
